Change file struction
