+++ src/main.rs
@@ -1,6 +1,6 @@
 use iced::Application;
-use rustic_rover::RusticRover;
-use rustic_rover::iced_utils::ApplicationDefaultSetting;
+use rustic_rover::core::RusticRover;
+use rustic_rover::core::iced_utils::ApplicationDefaultSetting;
 
 fn main()->iced::Result{
     let defalt_set = ApplicationDefaultSetting::new();
